% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{convergence}
\alias{convergence}
\title{R-hat convergence statistic (Documentation incomplete)}
\usage{
convergence(output, burn, n_eff = FALSE, graph = TRUE, verbose = TRUE)
}
\arguments{
\item{output}{A list outputted by \code{est_rbsvar()}.}

\item{burn}{...}

\item{n_eff}{...}

\item{graph}{...}

\item{verbose}{...}
}
\value{
A matrix of element-wise convergence statistics in the first column.
}
\description{
R-hat convergence statistic (Documentation incomplete)
}
