% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{eval_rbsvar}
\alias{eval_rbsvar}
\title{C++ wrapper for evaluating the posterior density (Documentation incomplete)}
\usage{
eval_rbsvar(model, par = NULL, parallel_likelihood = FALSE, max_cores = NA)
}
\arguments{
\item{model}{A list outputted by \code{init_rbsvar()}.}

\item{par}{...}

\item{parallel_likelihood}{...}

\item{max_cores}{...}
}
\value{
Proportional posterior density at the given point.
}
\description{
C++ wrapper for evaluating the posterior density (Documentation incomplete)
}
