% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{init_rbsvar}
\alias{init_rbsvar}
\title{Initializes rbsvar model (Documentation incomplete)}
\usage{
init_rbsvar(
  y,
  lags = 1,
  constant = TRUE,
  type = c("svar", "var")[1],
  garch = FALSE,
  mean_cent = TRUE,
  var_adj = FALSE,
  p_prior = c(2, 1),
  q_prior = c(1, 1),
  r_prior = c(log(1), 1),
  B_prior = c(NA, NA),
  B_inverse = TRUE,
  shrinkage = Inf,
  minnesota_means = NULL,
  prior = list(),
  init = NULL,
  maxit = 0,
  method = "L-BFGS-B",
  skip_hessian = NULL,
  trace = 1,
  REPORT = 100,
  min_density = -1e+06,
  parallel_likelihood = FALSE,
  max_cores = NA,
  verbose = TRUE
)
}
\arguments{
\item{y}{The data.}

\item{lags}{...}

\item{constant}{...}

\item{type}{...}

\item{garch}{...}

\item{mean_cent}{...}

\item{var_adj}{...}

\item{p_prior}{...}

\item{q_prior}{...}

\item{r_prior}{...}

\item{B_prior}{...}

\item{B_inverse}{...}

\item{shrinkage}{...}

\item{minnesota_means}{...}

\item{prior}{...}

\item{init}{...}

\item{maxit}{...}

\item{method}{...}

\item{skip_hessian}{...}

\item{trace}{...}

\item{REPORT}{...}

\item{min_density}{...}

\item{parallel_likelihood}{...}

\item{max_cores}{...}

\item{verbose}{...}
}
\value{
A list that may be passed for example to \code{est_rbsvar()} or \code{eval_rbsvar()}.
}
\description{
Initializes rbsvar model (Documentation incomplete)
}
