% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{irf}
\alias{irf}
\title{Computes impulse response functions (Documentation incomplete)}
\usage{
irf(
  model,
  output,
  burn = 0,
  horizon = 40,
  N = 1000,
  cumulate = c(),
  shock_sizes = NULL,
  shocks = NULL,
  verbose = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{model}{A list outputted by \code{init_rbsvar()}.}

\item{output}{A list outputted by \code{est_rbsvar()}.}

\item{burn}{...}

\item{horizon}{...}

\item{N}{...}

\item{cumulate}{...}

\item{shock_sizes}{...}

\item{shocks}{...}

\item{verbose}{...}

\item{parallel}{...}
}
\value{
A list that may be passed to \code{irf_plot()}.
}
\description{
Computes impulse response functions (Documentation incomplete)
}
