% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{irf_plot}
\alias{irf_plot}
\title{Plots impulse response functions (Documentation incomplete)}
\usage{
irf_plot(
  irf_obj,
  varnames = NULL,
  probs = c(0.05, 0.16, 0.84, 0.95),
  mar = c(2, 2, 2, 1),
  mfrow = NULL,
  color = "tomato",
  leg = TRUE,
  plot_median = FALSE,
  normalize = NULL
)
}
\arguments{
\item{irf_obj}{A list outputted by \code{irf()}.}

\item{varnames}{...}

\item{probs}{...}

\item{mar}{...}

\item{mfrow}{...}

\item{color}{...}

\item{leg}{...}

\item{plot_median}{...}

\item{normalize}{...}
}
\description{
Plots impulse response functions (Documentation incomplete)
}
