% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{marginal_likelihood}
\alias{marginal_likelihood}
\title{Computes a numerical estimate for the log-marginal-likelihood (Documentation incomplete)}
\usage{
marginal_likelihood(
  output,
  model,
  burn,
  M = NA,
  J = NULL,
  rel_tune = 1,
  parallel = FALSE,
  parallel_likelihood = FALSE
)
}
\arguments{
\item{output}{A list outputted by \code{est_rbsvar()}.}

\item{model}{A list outputted by \code{init_rbsvar()}.}

\item{burn}{...}

\item{M}{...}

\item{J}{...}

\item{rel_tune}{...}

\item{parallel}{...}

\item{parallel_likelihood}{...}
}
\value{
A list.
}
\description{
Computes a numerical estimate for the log-marginal-likelihood (Documentation incomplete)
}
