% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{plot_chains}
\alias{plot_chains}
\title{Plots the chains and marginal distirbution w.r.t one parameter (Documentation incomplete)}
\usage{
plot_chains(
  output,
  variable = "b1",
  burn = 0,
  breaks = "Sturges",
  real_val = NA,
  n_eff = FALSE
)
}
\arguments{
\item{output}{A list outputted by \code{est_rbsvar()}.}

\item{variable}{...}

\item{burn}{...}

\item{breaks}{...}

\item{real_val}{...}

\item{n_eff}{...}
}
\description{
Plots the chains and marginal distirbution w.r.t one parameter (Documentation incomplete)
}
