% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{plot_densities}
\alias{plot_densities}
\title{Plots the densities of the chains as a function of iterations (Documentation incomplete)}
\usage{
plot_densities(output, verbose = TRUE, drop = 1)
}
\arguments{
\item{output}{A list outputted by \code{est_rbsvar()}.}

\item{verbose}{...}

\item{drop}{...}
}
\description{
Plots the densities of the chains as a function of iterations (Documentation incomplete)
}
