% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{post_sample}
\alias{post_sample}
\title{Convenience function returning a sample from the posterior (Documentation incomplete)}
\usage{
post_sample(output, burn = 0, N = NA)
}
\arguments{
\item{output}{A list outputted by \code{est_rbsvar()}.}

\item{burn}{...}

\item{N}{...}
}
\value{
A numerical matrix.
}
\description{
Convenience function returning a sample from the posterior (Documentation incomplete)
}
