% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{sign_probabilities}
\alias{sign_probabilities}
\title{Computes sign restriction probabilities a'la Lanne & Luoto (2020) (Documentation incomplete)}
\usage{
sign_probabilities(
  model,
  output,
  burn,
  signs,
  horizons,
  limit = 3.2,
  N = 10000,
  cumulate = c(),
  verbose = TRUE
)
}
\arguments{
\item{model}{A list outputted by \code{init_rbsvar()}.}

\item{output}{A list outputted by \code{est_rbsvar()}.}

\item{burn}{...}

\item{signs}{...}

\item{horizons}{...}

\item{limit}{...}

\item{N}{...}

\item{cumulate}{...}

\item{verbose}{...}
}
\value{
A list.
}
\description{
Computes sign restriction probabilities a'la Lanne & Luoto (2020) (Documentation incomplete)
}
